inline memory_safety_test
