+++ rs/tests/message_routing/BUILD.bazel
@@ -3,6 +3,10 @@ load("//rs/tests:system_tests.bzl", "system_test", "system_test_nns")
 
 package(default_visibility = ["//rs:system-tests-pkg"])
 
+exports_files([
+    "memory_safety.wasm",
+])
+
 system_test_nns(
     name = "global_reboot_test",
     env = {
@@ -28,14 +32,29 @@ system_test_nns(
 
 system_test(
     name = "memory_safety_test",
-    proc_macro_deps = MACRO_DEPENDENCIES,
+    compile_data = [
+        "memory_safety.wasm",
+    ],
     tags = [
         "k8s",
         "system_test_hourly",
     ],
     target_compatible_with = ["@platforms//os:linux"],  # requires libssh that does not build on Mac OS
     runtime_deps = GUESTOS_RUNTIME_DEPS,
-    deps = DEPENDENCIES + ["//rs/tests"],
+    deps = [
+        "//rs/registry/subnet_type",
+        "//rs/tests/driver:ic-system-test-driver",
+        "//rs/types/base_types",
+        "//rs/types/types",
+        "@crate_index//:anyhow",
+        "@crate_index//:candid",
+        "@crate_index//:ic-agent",
+        "@crate_index//:ic-utils",
+        "@crate_index//:rand",
+        "@crate_index//:rand_chacha",
+        "@crate_index//:slog",
+        "@crate_index//:tokio",
+    ],
 )
 
 system_test_nns(
