+++ rs/tests/message_routing/memory_safety.wasm
